requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
dj-database-url==1.3.0
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-password-validators==1.7.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.2.0
